Move getStringAsJsonArray to JsonUtils
