+++ app/src/main/java/org/ole/planet/myplanet/ui/exam/ExamAnswerUtils.kt
@@ -3,7 +3,7 @@ package org.ole.planet.myplanet.ui.exam
 import java.util.Arrays
 import java.util.Locale
 import org.ole.planet.myplanet.model.RealmExamQuestion
-import org.ole.planet.myplanet.utilities.JsonParserUtils.getStringAsJsonArray
+import org.ole.planet.myplanet.utilities.JsonUtils.getStringAsJsonArray
 
 object ExamAnswerUtils {
     fun getChoiceTextById(question: RealmExamQuestion, id: String): String {
